weinc_src/front-store-main/requirements.txt
aioredis==2.0.1
anyio==3.6.2
astroid==2.12.13
async-timeout==4.0.3
bcrypt==4.0.1
bleach==6.0.0
boto3==1.27.0
botocore==1.30.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==40.0.2
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.87.0
fastapi-cache2==0.2.1
fastapi-redis-cache==0.2.5
Flask==2.2.3
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httptools==0.5.0
httpx==0.25.1
idna==3.4
importlib-metadata==6.0.0
inflect==6.0.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.8.0
MarkupSafe==2.1.2
mccabe==0.7.0
pendulum==3.0.0
platformdirs==2.5.4
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pylint==2.15.6
pylint-pydantic==0.1.4
pymongo==4.3.3
PyMySQL==1.0.2
pyOpenSSL==23.1.1
pyotp==2.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python-telegram-bot==20.6
PyYAML==6.0
redis==4.6.0
requests==2.28.2
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
time-machine==2.14.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2024.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
